Breast Cancer Tracker Ready.
Integrated Model Save in the training helper function.  I learned that

Added same transformtiaon to the Test features as it was added to the Train Features. I thought it was good practice not to add transformation to the Test Features, but I think that only applies to Image Classification not tabular datas.